pyproject.toml
[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |\.tox
  |Dockerfile
  |Jenkinfile
  | _build
  | dist
)/
'''

# ==== isort ====
[tool.isort]
profile = "black"
line_length = 120
known_first_party = [
    "config",
    "apps",
]
skip = []
skip_glob = []

[tool.poetry]
name = "simplelinkedin"
version = "1.0.0"
description = "Python package to automate activities on LinkedIn."
authors = ["Vishal Kumar Mishra <vishal.k.mishra2@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/TheConfused/LinkedIn"
repository = "https://github.com/TheConfused/LinkedIn"
keywords = ["linkedin", "simplelinkedin", "selenium", "linkedin automation"]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.12.0"
simpleselenium = "^0.4.1"
django-environ = "^0.11.1"
python-dotenv = "^1.0.0"
PyYAML = "^6.0.1"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings --reuse-db"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== mypy ====
[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true


[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 120
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.bandit]
exclude_dirs = [".tox"]
# tests = ["B201", "B301"]
skips = ["B113"]
